Changed to a new more advanced pos list
Changed the idea of a word to encapsulate one inflection of a word such
as "went" or "gone" as opposed to the set of words that come from one
root word.
Switched to the new list, which required many changes.

Began the changes to the decoding of the file in order to get all the
information out of it. This is all going into the utility class
PartsOfSpeech